Import sourceforge tarball: abills-0.50.tgz

+++ Abills/mysql/Dv_Sessions.pm
@@ -50,11 +50,24 @@ sub del {
 
   if ($attr->{DELETE_USER}) {
     $self->query($db, "DELETE FROM dv_log WHERE uid='$attr->{DELETE_USER}';", 'do');
-  }
-  else {
-    $self->query($db, "DELETE FROM dv_log 
-     WHERE uid='$uid' and start='$session_start' and nas_id='$nas_id' and acct_session_id='$session_id';", 'do');
    }
+  else {
+  	$self->query($db, "show tables LIKE 'traffic_prepaid_sum'");
+  	
+  	if ($self->{TOTAL} > 0) {
+      $self->query($db, "UPDATE traffic_prepaid_sum pl, dv_log l SET 
+         traffic_in=traffic_in-(l.recv + 4294967296 * acct_input_gigawords),
+         traffic_out=traffic_out-(l.sent + 4294967296 * acct_output_gigawords)
+         WHERE pl.uid=l.uid AND l.uid='$uid' and l.start='$session_start' and l.nas_id='$nas_id' 
+          and l.acct_session_id='$session_id';", 'do');
+  	 }
+  	
+     $self->query($db, "DELETE FROM dv_log 
+       WHERE uid='$uid' and start='$session_start' and nas_id='$nas_id' and acct_session_id='$session_id';", 'do');
+   }
+
+   
+
 
   return $self;
 }
@@ -143,7 +156,8 @@ sub online {
 	my ($attr) = @_;
 
   my $WHERE = '';
- 
+
+  $admin->{DOMAIN_ID}=0 if (! $admin->{DOMAIN_ID});
 
   if ($attr->{COUNT}) {
   	if ($attr->{ZAPED}) {
@@ -235,7 +249,9 @@ sub online {
    ACCT_SESSION_TIME => 'UNIX_TIMESTAMP() - UNIX_TIMESTAMP(c.started)',
    DURATION_SEC    => 'c.lupdated - UNIX_TIMESTAMP(c.started)',
    FILTER_ID       => 'dv.filter_id',
-   SESSION_START   => 'UNIX_TIMESTAMP(started)'
+   SESSION_START   => 'UNIX_TIMESTAMP(started)',
+   DISABLE         => 'u.disable',
+   DV_STATUS       => 'dv.disable'
   );
 
 
@@ -440,6 +456,11 @@ sub online_del {
     $WHERE = "nas_id='$NAS_ID'
             and nas_port_id='$NAS_PORT' 
             and acct_session_id='$ACCT_SESSION_ID'";
+   }
+
+  $self->query($db, "SELECT uid, user_name FROM dv_calls WHERE $WHERE");
+  foreach my $line ( @{  $self->{list} } ) {
+    $admin->action_add("$line->[0]", "$line->[1]", { MODULE => 'Dv', TYPE => 13 });
    }
 
   $self->query($db, "DELETE FROM dv_calls WHERE $WHERE;", 'do');
